fix: create commands/normal.ts
