+++ src/content/App.tsx
@@ -46,121 +46,6 @@ const App: React.FC = () => {
     const { lines, charCount, currentLine, col } = getLines(element, start);
     const { currentLine: endCurrentLine } = getLines(element, end);
 
-    if (
-      mode.current === "normal" &&
-      !matchesShortcut(e, shortcuts.normal_mode)
-    ) {
-      if (e.key === "h" && col) {
-        start--;
-        end = start + 1;
-      }
-
-      if (e.key === "l" && col !== lines[currentLine].length - 1) {
-        start++;
-        end = start + 1;
-      }
-
-      if (e.key === "j" && currentLine + 1 < lines.length) {
-        const nextLineLength = lines[currentLine + 1].length;
-        const next = charCount + lines[currentLine].length + 1;
-
-        if (nextLineLength) {
-          const overCharCount =
-            col >= nextLineLength ? col - nextLineLength + 1 : 0;
-          start = next + col - overCharCount;
-        } else {
-          start = next;
-        }
-        end = start + 1;
-      }
-
-      if (e.key === "k" && currentLine > 0) {
-        const prevLineLength = lines[currentLine - 1].length;
-        const prev = charCount - (lines[currentLine - 1].length + 1);
-
-        if (prevLineLength) {
-          const overCharCount =
-            col >= prevLineLength ? col - prevLineLength + 1 : 0;
-          start = prev + col - overCharCount;
-        } else {
-          start = prev;
-        }
-        end = start + 1;
-      }
-
-      if (e.key === "J") {
-        const nextBreak = element.value.indexOf("\n", start);
-        if (nextBreak >= 0) {
-          start = nextBreak;
-          end = nextBreak + 1;
-          insertText(element, start, end, " ");
-        }
-      }
-
-      if (e.key === "o" && element.tagName === "TEXTAREA") {
-        const nextBreak = element.value.indexOf("\n", start);
-        start = nextBreak === -1 ? length : nextBreak;
-        start = end = start + 1;
-        insertText(element, start, end, "\n");
-        mode.current = "insert";
-      }
-
-      if (e.key === "O" && element.tagName === "TEXTAREA") {
-        const prevBreak = element.value.lastIndexOf("\n", start - 1);
-        start = prevBreak === -1 ? 0 : prevBreak;
-        start = end = start + (start ? 1 : 0);
-        insertText(element, start, end, "\n");
-        mode.current = "insert";
-      }
-
-      if (e.key === "p") {
-        const text = await navigator.clipboard.readText();
-        if (text) {
-          const pos = lines[currentLine].length === 0 ? start : end;
-          insertText(element, pos, pos, text);
-          start = start + text.length;
-          end = start + 1;
-        }
-      }
-
-      if (e.key === "P") {
-        const text = await navigator.clipboard.readText();
-        if (text) {
-          insertText(element, start, start, text);
-          start = start + text.length - 1;
-          end = start + 1;
-        }
-      }
-
-      if (e.key === "x") {
-        const text = window.getSelection()?.toString();
-        if (text) {
-          navigator.clipboard.writeText(text);
-          insertText(element, start, end, "");
-        }
-      }
-
-      if (e.key === "X") {
-        start = start - 1;
-        end = end - 1;
-        navigator.clipboard.writeText(element.value.slice(start, start + 1));
-        insertText(element, start, end, "");
-      }
-
-      if (lines[currentLine].length && col === lines[currentLine].length) {
-        start = start - 1;
-        end = start + 1;
-      }
-
-      if (e.key === "u") {
-        document.execCommand("undo");
-      }
-
-      if (e.key === "r" && e.ctrlKey) {
-        document.execCommand("redo");
-      }
-    }
-
     if (mode.current === "visual" && originalPos.current) {
       const {
         start: oStart,
@@ -312,6 +197,121 @@ const App: React.FC = () => {
     pos.current = { start, end };
     element.setSelectionRange(start, end);
   };
+  //   if (
+  //     mode.current === "normal" &&
+  //     !matchesShortcut(e, shortcuts.normal_mode)
+  //   ) {
+  //     if (e.key === "h" && col) {
+  //       start--;
+  //       end = start + 1;
+  //     }
+  //
+  //     if (e.key === "l" && col !== lines[currentLine].length - 1) {
+  //       start++;
+  //       end = start + 1;
+  //     }
+  //
+  //     if (e.key === "j" && currentLine + 1 < lines.length) {
+  //       const nextLineLength = lines[currentLine + 1].length;
+  //       const next = charCount + lines[currentLine].length + 1;
+  //
+  //       if (nextLineLength) {
+  //         const overCharCount =
+  //           col >= nextLineLength ? col - nextLineLength + 1 : 0;
+  //         start = next + col - overCharCount;
+  //       } else {
+  //         start = next;
+  //       }
+  //       end = start + 1;
+  //     }
+  //
+  //     if (e.key === "k" && currentLine > 0) {
+  //       const prevLineLength = lines[currentLine - 1].length;
+  //       const prev = charCount - (lines[currentLine - 1].length + 1);
+  //
+  //       if (prevLineLength) {
+  //         const overCharCount =
+  //           col >= prevLineLength ? col - prevLineLength + 1 : 0;
+  //         start = prev + col - overCharCount;
+  //       } else {
+  //         start = prev;
+  //       }
+  //       end = start + 1;
+  //     }
+  //
+  //     if (e.key === "J") {
+  //       const nextBreak = element.value.indexOf("\n", start);
+  //       if (nextBreak >= 0) {
+  //         start = nextBreak;
+  //         end = nextBreak + 1;
+  //         insertText(element, start, end, " ");
+  //       }
+  //     }
+  //
+  //     if (e.key === "o" && element.tagName === "TEXTAREA") {
+  //       const nextBreak = element.value.indexOf("\n", start);
+  //       start = nextBreak === -1 ? length : nextBreak;
+  //       start = end = start + 1;
+  //       insertText(element, start, end, "\n");
+  //       mode.current = "insert";
+  //     }
+  //
+  //     if (e.key === "O" && element.tagName === "TEXTAREA") {
+  //       const prevBreak = element.value.lastIndexOf("\n", start - 1);
+  //       start = prevBreak === -1 ? 0 : prevBreak;
+  //       start = end = start + (start ? 1 : 0);
+  //       insertText(element, start, end, "\n");
+  //       mode.current = "insert";
+  //     }
+  //
+  //     if (e.key === "p") {
+  //       const text = await navigator.clipboard.readText();
+  //       if (text) {
+  //         const pos = lines[currentLine].length === 0 ? start : end;
+  //         insertText(element, pos, pos, text);
+  //         start = start + text.length;
+  //         end = start + 1;
+  //       }
+  //     }
+  //
+  //     if (e.key === "P") {
+  //       const text = await navigator.clipboard.readText();
+  //       if (text) {
+  //         insertText(element, start, start, text);
+  //         start = start + text.length - 1;
+  //         end = start + 1;
+  //       }
+  //     }
+  //
+  //     if (e.key === "x") {
+  //       const text = window.getSelection()?.toString();
+  //       if (text) {
+  //         navigator.clipboard.writeText(text);
+  //         insertText(element, start, end, "");
+  //       }
+  //     }
+  //
+  //     if (e.key === "X") {
+  //       start = start - 1;
+  //       end = end - 1;
+  //       navigator.clipboard.writeText(element.value.slice(start, start + 1));
+  //       insertText(element, start, end, "");
+  //     }
+  //
+  //     if (lines[currentLine].length && col === lines[currentLine].length) {
+  //       start = start - 1;
+  //       end = start + 1;
+  //     }
+  //
+  //     if (e.key === "u") {
+  //       document.execCommand("undo");
+  //     }
+  //
+  //     if (e.key === "r" && e.ctrlKey) {
+  //       document.execCommand("redo");
+  //     }
+  //   }
+  //
 
   useEffect(() => {
     loadShortcuts().then(setShortcuts);
